refactor(settings): refactored the settings component into reusable components
re #1479
