+++ src/locales/en.json
@@ -82,6 +82,7 @@
   "Registering decentralized identity": "Registering decentralized identity",
   "Restore account": "Restore account",
   "Registration failed": "Registration failed",
+  "Security": "Security",
   "Self-signed": "Self-signed",
   "Settings": "Settings",
   "Set up a secure phrase to recover your account in the future if your phone is stolen or is damaged": "Set up a secure phrase to recover your account in the future if your phone is stolen or is damaged",
